LayoutTests/http/tests/cookies/same-site/popup-same-site-post-expected.txt


PASS '127.0.0.1' is same-site with itself, so samesite cookies are sent via POST.
